requirements.txt
apiview==1.3.25
asgiref==3.6.0
bunch==1.0.1
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
configparser==5.3.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.1
cssselect==1.2.0
cssutils==2.7.1
defusedxml==0.7.1
Deprecated==1.2.14
diff-match-patch==20200713
Django==4.2
django-braces==1.15.0
django-cors-headers==4.0.0
django-dotenv==1.4.2
django-filter==23.2
django-import-export==3.2.0
django-jazzmin==2.6.0
django-oauth-toolkit==2.3.0
django-rest-auth==0.9.5
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-social-oauth2==2.1.3
drf-writable-nested==0.7.0
drf-yasg==1.21.5
ecdsa==0.18.0
et-xmlfile==1.1.0
idna==3.4
inflection==0.5.1
install==1.3.5
itypes==1.2.0
Jinja2==3.1.2
jwcrypto==1.5.0
lxml==4.9.2
MarkupPy==1.14
MarkupSafe==2.1.2
mock==5.0.2
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.1
Pillow==9.5.0
premailer==3.10.0
psycopg2-binary==2.9.6
pyasn1==0.5.0
pycparser==2.21
PyJWT==2.7.0
python-jose==3.3.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
requests==2.30.0
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
simple-configparser==0.1.4.0
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
tablib==3.4.0
templated-mail==0.4.0.2
typing_extensions==4.6.3
uritemplate==4.1.1
urllib3==2.0.2
wrapt==1.15.0
xlrd==2.0.1
xlwt==1.3.0
yagmail==0.15.293
